# openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/native-auth/start": {
            "post": {
                "operationId": "startNativeAuth",
                "parameters": [{ "name": "origin", "required": true, "in": "header", "schema": { "type": "string" } }],
                "responses": {
                    "201": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "flow": { "type": "string", "pattern": "^[a-f0-9]{64}$" },
                                        "browserUrl": { "type": "string", "format": "uri" },
                                        "expiresIn": {
                                            "type": "integer",
                                            "minimum": 0,
                                            "exclusiveMinimum": true,
                                            "maximum": 9007199254740991
                                        }
                                    },
                                    "required": ["flow", "browserUrl", "expiresIn"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Stable API error with a request ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "minimum": -9007199254740991,
                                            "maximum": 9007199254740991
                                        },
                                        "code": { "type": "string" },
                                        "message": { "type": "string" },
                                        "requestId": { "type": "string" },
                                        "details": { "type": "array", "items": { "type": "string" } }
                                    },
                                    "required": ["statusCode", "code", "message", "requestId"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    }
                },
                "tags": ["native-auth"],
                "summary": "Begin native sign-in or password reset",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "intent": {
                                        "default": "sign-in",
                                        "type": "string",
                                        "enum": ["sign-in", "password-reset"]
                                    },
                                    "challenge": { "type": "string", "pattern": "^[A-Za-z0-9_-]{43}$" },
                                    "state": { "type": "string", "pattern": "^[A-Za-z0-9_-]{43}$" }
                                },
                                "required": ["challenge", "state"],
                                "additionalProperties": false
                            }
                        }
                    }
                }
            }
        },
        "/api/native-auth/authorize": {
            "post": {
                "operationId": "authorizeNativeAuth",
                "parameters": [{ "name": "origin", "required": true, "in": "header", "schema": { "type": "string" } }],
                "responses": {
                    "201": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": { "callbackUrl": { "type": "string", "format": "uri" } },
                                    "required": ["callbackUrl"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Stable API error with a request ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "minimum": -9007199254740991,
                                            "maximum": 9007199254740991
                                        },
                                        "code": { "type": "string" },
                                        "message": { "type": "string" },
                                        "requestId": { "type": "string" },
                                        "details": { "type": "array", "items": { "type": "string" } }
                                    },
                                    "required": ["statusCode", "code", "message", "requestId"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Stable API error with a request ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "minimum": -9007199254740991,
                                            "maximum": 9007199254740991
                                        },
                                        "code": { "type": "string" },
                                        "message": { "type": "string" },
                                        "requestId": { "type": "string" },
                                        "details": { "type": "array", "items": { "type": "string" } }
                                    },
                                    "required": ["statusCode", "code", "message", "requestId"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    }
                },
                "tags": ["native-auth"],
                "summary": "Approve native sign-in from a browser session",
                "security": [{ "session": [] }, { "nativeToken": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": { "flow": { "type": "string" } },
                                "required": ["flow"]
                            }
                        }
                    }
                }
            }
        },
        "/api/native-auth/exchange": {
            "post": {
                "operationId": "exchangeNativeAuth",
                "parameters": [{ "name": "origin", "required": true, "in": "header", "schema": { "type": "string" } }],
                "responses": {
                    "201": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": { "type": "string" },
                                        "resetToken": { "type": "string" },
                                        "expiresAt": {
                                            "type": "string",
                                            "format": "date-time",
                                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                                        }
                                    },
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Stable API error with a request ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "minimum": -9007199254740991,
                                            "maximum": 9007199254740991
                                        },
                                        "code": { "type": "string" },
                                        "message": { "type": "string" },
                                        "requestId": { "type": "string" },
                                        "details": { "type": "array", "items": { "type": "string" } }
                                    },
                                    "required": ["statusCode", "code", "message", "requestId"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    }
                },
                "tags": ["native-auth"],
                "summary": "Exchange a native authorization code",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "flow": { "type": "string", "pattern": "^[a-f0-9]{64}$" },
                                    "code": { "type": "string", "pattern": "^[a-f0-9]{64}$" },
                                    "verifier": { "type": "string", "pattern": "^[A-Za-z0-9_-]{43,128}$" },
                                    "state": { "type": "string", "pattern": "^[A-Za-z0-9_-]{43}$" }
                                },
                                "required": ["flow", "code", "verifier", "state"],
                                "additionalProperties": false
                            }
                        }
                    }
                }
            }
        },
        "/api/native-auth/reset-handoff": {
            "post": {
                "operationId": "resetNativeAuth",
                "parameters": [{ "name": "origin", "required": true, "in": "header", "schema": { "type": "string" } }],
                "responses": {
                    "201": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": { "callbackUrl": { "type": "string", "format": "uri" } },
                                    "required": ["callbackUrl"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Stable API error with a request ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "minimum": -9007199254740991,
                                            "maximum": 9007199254740991
                                        },
                                        "code": { "type": "string" },
                                        "message": { "type": "string" },
                                        "requestId": { "type": "string" },
                                        "details": { "type": "array", "items": { "type": "string" } }
                                    },
                                    "required": ["statusCode", "code", "message", "requestId"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    }
                },
                "tags": ["native-auth"],
                "summary": "Hand off a password reset to a native app",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": { "flow": { "type": "string" }, "token": { "type": "string" } },
                                "required": ["flow", "token"]
                            }
                        }
                    }
                }
            }
        },
...
    },
    "info": { "title": "Fullstack application API", "description": "", "version": "1.0.0", "contact": {} },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "session": {
                "type": "apiKey",
                "in": "cookie",
                "name": "better-auth.session_token",
                "description": "Browser session; native clients use bearer tokens."
            },
            "nativeToken": { "type": "http", "scheme": "bearer", "description": "Native secure-storage session token." }
        },
...
    }
}